feat: testing, stok telur

+++ dto/Respon/EggAvailableRespon.cs
@@ -4,14 +4,14 @@ namespace sipetok_api.Respon
 {
     public class EggAvailableRespon
     {
-        public int stock { get; set; }
+        public double stock { get; set; }
         public int tenant_id { get; set; }
         public int category_id { get; set; }
         public virtual EggCategoryRespon? category { get; set; }
 
         public EggAvailableRespon() { }
 
-        public EggAvailableRespon(int stock, int tenant_id, int category_id)
+        public EggAvailableRespon(double stock, int tenant_id, int category_id)
         {
             this.stock = stock;
             this.tenant_id = tenant_id;
